requirements.txt
requests
markdownify
openai>=1.0.0
python-dotenv
streamlit
chromadb
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-chroma